refactor(document menu): rename and move
Was called `preview-menu`.
