+++ web/src/ui/document/menu.ts
@@ -2,8 +2,9 @@ import { apply } from '@twind/core'
 import { LitElement, html } from 'lit'
 import { customElement, property, state } from 'lit/decorators'
 
-import { NodeChipState } from '../../contexts/preview-context'
 import { withTwind } from '../../twind'
+
+import { NodeChipState } from './context'
 
 import '../buttons/icon'
